chore: clippy --all-targets sweep — tests lint-clean under -D warnings
Every gate now runs against all targets (lib, bins, tests, benches):
cargo clippy --workspace --all-targets -- -D warnings, workspace tests
(47 suites), fmt --check, deny, kernel-purity, fingerprint, gen-schemas,
and both SLO benches all green.

+++ crates/exocortex-dreams/src/lib.rs
@@ -259,7 +259,7 @@ impl<S: Storage + 'static> DreamsEngine<S> {
         drop(ms);
         let now = chrono::Utc::now();
         rows.sort_by(|a, b| {
-            let score = |t: chrono::DateTime<chrono::Utc>| -(now - t).num_days() as i64;
+            let score = |t: chrono::DateTime<chrono::Utc>| -(now - t).num_days();
             score(b.0).cmp(&score(a.0)).then(b.1.id.cmp(&a.1.id))
         });
         Ok(rows.into_iter().take(32).map(|(_, m)| m).collect())
